Fix problem with bound properties not updating correctly.
I was seeing this crop up when using the component like so:

```handlebars
{{ivy-codemirror mode=mode}}
```

I think what's happening is that `ivy-codemirror` uses
`notifyPropertyChange` when setting up its option observers so that it
can synchronously fire them to set the initial values. The problem is
that `notifyPropertyChange` appears to have overwritten the `mode`
property on the model that it was bound to. The fix is to call the
observer functions directly rather than using `notifyPropertyChange`.