Ensure `valueUpdated` is in the `actions` queue
Previously we would send `valueUpdated` synchronously, but this can
cause a deprecation warning because it is sent during
`didInsertElement`, and Ember complains if you change component
properties during that time. Scheduling the action to be sent once, in
the `actions` queue, solves this problem. Note that I also had to
extract a separate method to do the actual work of calling `sendAction`,
since the action name is an argument, and `run.once` does not consider
arguments when looking for duplicates.